font: add a font builder
This takes one or more .hex files with font data and produces binary
blobs with single-wide and double-wide characters, as well a character
map. These are produced in two variants: one with the full set of
supported characters and a second one with a stripped down set for parts
with limited flash space.