Add a rowbuffer with test pattern
Create a new videoout module that holds all the logic for video output
including requesting the VGA resource and compositing the cursor. Also
adds a rowbuffer. Reading from the rowbuffer takes a cycle so hsync,
vsync, and den are delayed by a clock on the way to the output. Also,
this fixes the phase of the clock.