Add pack_data.py
This is a utility that allocates flash blocks such that addresses can be
computed by bit concatenation rather than addition. It also can flash
the appropriate blobs to the appropriate places in the flash of the
target device. For UPDuino it does this by repeatedly calling iceprog
whereas for TinyFPGA it writes all the data into a single blob which can
be flashed by a single tinyprog call, since tinyprog boots the device
after programming.