Move glyphbuffer into its own module
This new module handles two things: the arbitration between reads and
writes to be compatible with the SPRAM on iCE40 UP parts, and the
address generation from a row/column address, which is now done by
concatenation.
Also add a character generation module to fill the glyphbuffer with a
test pattern, since it can't necessarily be initialized the conventional
way anymore.